Fixed issue with PUNICODE_STRING usage
Passing a non-null terminated string to WinHTTP does not appear to
be supported. In some cases, this would result in the password not
being sent to the password checking service.

To resolve this issue, I have made copies of the password and
username in the form of wstrings. I use c_str() to pass the null
terminated string to WinHTTP.

I'm pretty certain that this is OK because I call SecureZeroMemory
on the wstring after its usage. The c_str() actually points to the
backing buffer for wstring (plus an additional character for the
null termination).

I would welcome any additional wisdom on this issue...