Modify code structure and logging
This commit modifies the structure of the code for checking passwords.
Rather than sloppily logging random print statements, the daemon now
outputs a concise log message summarizing the check results.

Additionally, this commit adds a hard-coded 3 second timeout for a
response from the HaveIBeenPwnd check.
This isn't a fantastic solution, but it moves towards a much better
implementation.